create endpoint to get venues
