+++ api/src/controllers/index.ts
@@ -2,6 +2,7 @@ import articlesController from "./articles";
 import articleController from "./article";
 import eventController from "./event";
 import eventsController from "./events";
+import venuesController from "./venue";
 import healthcheckController from "./healthcheck";
 
 export { errorHandler } from "./error";
@@ -10,5 +11,6 @@ export {
   articleController,
   eventController,
   eventsController,
+  venuesController,
   healthcheckController,
 };
